Add exception handler
