+++ handlers/src/test/kotlin/com/hexagonkt/handlers/HandlersTest.kt
@@ -1,0 +1,18 @@
+package com.hexagonkt.handlers
+
+import org.junit.jupiter.api.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertFailsWith
+
+internal class HandlersTest {
+
+    @Test fun `Exceptions are casted properly`() {
+        assertFailsWith<IllegalStateException> { castException(null, Exception::class) }
+        assertFailsWith<ClassCastException> {
+            castException(IllegalStateException(), IllegalArgumentException::class)
+        }
+
+        val ise = IllegalStateException()
+        assertEquals(ise, castException(ise, RuntimeException::class))
+    }
+}
